+++ scripts/demos/sensors/pva_sensor.py
@@ -10,7 +10,7 @@ import argparse
 from isaaclab.app import AppLauncher
 
 # add argparse arguments
-parser = argparse.ArgumentParser(description="Example on using the IMU sensor.")
+parser = argparse.ArgumentParser(description="Example on using the PVA sensor.")
 parser.add_argument("--num_envs", type=int, default=1, help="Number of environments to spawn.")
 # append AppLauncher cli args
 AppLauncher.add_app_launcher_args(parser)
@@ -31,7 +31,7 @@ import warp as wp
 import isaaclab.sim as sim_utils
 from isaaclab.assets import AssetBaseCfg
 from isaaclab.scene import InteractiveScene, InteractiveSceneCfg
-from isaaclab.sensors import ImuCfg
+from isaaclab.sensors import PvaCfg
 from isaaclab.utils import configclass
 
 ##
@@ -41,8 +41,8 @@ from isaaclab_assets.robots.anymal import ANYMAL_C_CFG  # isort: skip
 
 
 @configclass
-class ImuSensorSceneCfg(InteractiveSceneCfg):
-    """Design the scene with IMU sensors on the robot."""
+class PvaSensorSceneCfg(InteractiveSceneCfg):
+    """Design the scene with sensors on the robot."""
 
     # ground plane
     ground = AssetBaseCfg(prim_path="/World/defaultGroundPlane", spawn=sim_utils.GroundPlaneCfg())
@@ -55,9 +55,9 @@ class ImuSensorSceneCfg(InteractiveSceneCfg):
     # robot
     robot = ANYMAL_C_CFG.replace(prim_path="{ENV_REGEX_NS}/Robot")
 
-    imu_RF = ImuCfg(prim_path="{ENV_REGEX_NS}/Robot/LF_FOOT")
+    pva_RF = PvaCfg(prim_path="{ENV_REGEX_NS}/Robot/LF_FOOT", debug_vis=True)
 
-    imu_LF = ImuCfg(prim_path="{ENV_REGEX_NS}/Robot/RF_FOOT")
+    pva_LF = PvaCfg(prim_path="{ENV_REGEX_NS}/Robot/RF_FOOT", debug_vis=True)
 
 
 def run_simulator(sim: sim_utils.SimulationContext, scene: InteractiveScene):
@@ -73,6 +73,9 @@ def run_simulator(sim: sim_utils.SimulationContext, scene: InteractiveScene):
             # reset counter
             count = 0
             # reset the scene entities
+            # root state
+            # we offset the root state by the origin since the states are written in simulation world frame
+            # if this is not done, then the robots will be spawned at the (0, 0, 0) of the simulation world
             root_pose = wp.to_torch(scene["robot"].data.default_root_pose).clone()
             root_pose[:, :3] += scene.env_origins
             scene["robot"].write_root_link_pose_to_sim_index(root_pose=root_pose)
@@ -90,8 +93,11 @@ def run_simulator(sim: sim_utils.SimulationContext, scene: InteractiveScene):
             scene.reset()
             print("[INFO]: Resetting robot state...")
         # Apply default actions to the robot
+        # -- generate actions/commands
         targets = wp.to_torch(scene["robot"].data.default_joint_pos)
+        # -- apply action to the robot
         scene["robot"].set_joint_position_target_index(target=targets)
+        # -- write data to sim
         scene.write_data_to_sim()
         # perform step
         sim.step()
@@ -103,13 +109,17 @@ def run_simulator(sim: sim_utils.SimulationContext, scene: InteractiveScene):
 
         # print information from the sensors
         print("-------------------------------")
-        print(scene["imu_LF"])
-        print("Received angular velocity: ", scene["imu_LF"].data.ang_vel_b)
-        print("Received linear acceleration: ", scene["imu_LF"].data.lin_acc_b)
+        print(scene["pva_LF"])
+        print("Received linear velocity: ", scene["pva_LF"].data.lin_vel_b)
+        print("Received angular velocity: ", scene["pva_LF"].data.ang_vel_b)
+        print("Received linear acceleration: ", scene["pva_LF"].data.lin_acc_b)
+        print("Received angular acceleration: ", scene["pva_LF"].data.ang_acc_b)
         print("-------------------------------")
-        print(scene["imu_RF"])
-        print("Received angular velocity: ", scene["imu_RF"].data.ang_vel_b)
-        print("Received linear acceleration: ", scene["imu_RF"].data.lin_acc_b)
+        print(scene["pva_RF"])
+        print("Received linear velocity: ", scene["pva_RF"].data.lin_vel_b)
+        print("Received angular velocity: ", scene["pva_RF"].data.ang_vel_b)
+        print("Received linear acceleration: ", scene["pva_RF"].data.lin_acc_b)
+        print("Received angular acceleration: ", scene["pva_RF"].data.ang_acc_b)
 
 
 def main():
@@ -121,7 +131,7 @@ def main():
     # Set main camera
     sim.set_camera_view(eye=[3.5, 3.5, 3.5], target=[0.0, 0.0, 0.0])
     # design scene
-    scene_cfg = ImuSensorSceneCfg(num_envs=args_cli.num_envs, env_spacing=2.0)
+    scene_cfg = PvaSensorSceneCfg(num_envs=args_cli.num_envs, env_spacing=2.0)
     scene = InteractiveScene(scene_cfg)
     # Play the simulator
     sim.reset()